# docs/swagger.yaml
basePath: /
definitions:
...
  httpapi.CreatePlatformParamDictRequest:
    properties:
      builtin:
        type: boolean
      description:
        type: string
      name:
        type: string
      param_key:
        type: string
      param_type:
        type: string
      required:
        type: boolean
      status:
        type: integer
    type: object
...
  httpapi.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  httpapi.PlatformParamDictDataResponse:
    properties:
      data:
        $ref: '#/definitions/httpapi.PlatformParamDictResponse'
    type: object
  httpapi.PlatformParamDictListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/httpapi.PlatformParamDictResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  httpapi.PlatformParamDictResponse:
    properties:
      builtin:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      param_key:
        type: string
      param_type:
        type: string
      required:
        type: boolean
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  httpapi.UpdatePlatformParamDictRequest:
    properties:
      builtin:
        type: boolean
      description:
        type: string
      name:
        type: string
      param_key:
        type: string
      param_type:
        type: string
      required:
        type: boolean
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: Internal deployment platform API.
  title: GOS API
  version: "1.0"
paths:
...
  /platform-param-dicts:
    get:
      parameters:
      - description: Platform param key
        in: query
        name: param_key
        type: string
      - description: Platform param name
        in: query
        name: name
        type: string
      - description: 'Status: 1 enabled, 0 disabled'
        in: query
        name: status
        type: integer
      - description: Builtin flag
        in: query
        name: builtin
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.PlatformParamDictListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: List platform param dicts
      tags:
      - platform-param-dicts
    post:
      consumes:
      - application/json
      parameters:
      - description: Create platform param dict request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.CreatePlatformParamDictRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.PlatformParamDictDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Create platform param dict
      tags:
      - platform-param-dicts
  /platform-param-dicts/{id}:
    delete:
      parameters:
      - description: Platform param dict ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Delete platform param dict
      tags:
      - platform-param-dicts
    get:
      parameters:
      - description: Platform param dict ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.PlatformParamDictDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Get platform param dict by ID
      tags:
      - platform-param-dicts
    put:
      consumes:
      - application/json
      parameters:
      - description: Platform param dict ID
        in: path
        name: id
        required: true
        type: string
      - description: Update platform param dict request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.UpdatePlatformParamDictRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.PlatformParamDictDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Update platform param dict
      tags:
      - platform-param-dicts
...
schemes:
- http
- https
swagger: "2.0"